requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
faiss-cpu
pypdf
python-docx
python-dotenv
gradio
fastapi
langserve
uvicorn
sentence-transformers